switch to float zero_point

+++ cpp/matmul_quant_cuda.h
@@ -5,4 +5,4 @@
 using namespace torch;
 
 Tensor matmul_quant_cuda(Tensor a, Tensor b);
-Tensor quantize_and_pack(Tensor a, unsigned int bits, double scale, int64_t zero_point);
+Tensor quantize_and_pack(Tensor a, unsigned int bits, float scale, float zero_point);
